Split pp.py into a package (pure move, no logic change)
pp.py had grown to 2100 lines spanning the registry, plot matching, data
wrangling, filtering and Altair chart construction. Split it into
salk_toolkit/pp/ along those seams so the follow-up optimization work is
reviewable.

This is a pure move: every one of the 49 top-level defs and classes is
byte-identical to its pp.py original (verified by AST extraction), and
pp/__init__.py re-exports the full previous import surface, so
'from salk_toolkit.pp import X' keeps working for every X that worked
before. Pyright reports the same 114 diagnostics as before the split and
the suite is unchanged at 288 passed / 1 skipped.
